requirements.txt
# # ======================================================
# # Core Python & Utilities
# # ======================================================
# numpy>=1.23,<2.0
# scipy>=1.9
# tqdm>=4.65
# matplotlib>=3.7
# pandas>=1.5

# # ======================================================
# # PyTorch setup via Conda (Not recommended, try to run the pip command given in SETUP.md)
# # ======================================================
# # torch>=2.1,<2.3
# # torchaudio>=2.1,<2.3
# # torchvision>=0.16,<0.18

# # ======================================================
# # Hugging Face – ASR Models
# # ======================================================
# transformers>=4.36,<4.41
# datasets>=2.14
# sentencepiece>=0.1.99
# soundfile>=0.12

# # ======================================================
# # ONNX & ONNX Runtime
# # ======================================================
# onnx>=1.14,<1.16
# onnxruntime>=1.16,<1.18
# onnxruntime-gpu>=1.16,<1.18

# # ======================================================
# # Evaluation
# # ======================================================
# jiwer>=3.0

# # ======================================================
# # Profiling & Utilities
# # ======================================================
# psutil>=5.9
# tabulate>=0.9

# # ======================================================
# # Optional (Visualization / Debugging)
# # ======================================================
# ipython>=8.12
# jupyterlab>=4.0

# #=======================================================
# # TensorRT Insallation for Linux
# #=======================================================
# tensorrt-cu12 
# tensorrt-lean-cu12 
# tensorrt-dispatch-cu12

# ======================================================
# Core Python & Scientific Stack
# ======================================================
numpy==1.25.2
scipy==1.11.1
pandas==2.2.2
tqdm==4.65.0
psutil==5.9.5
tabulate==0.9.0

# ======================================================
# Plotting & Visualization
# ======================================================
matplotlib==3.7.2
seaborn==0.12.2

# ======================================================
# Audio Processing
# ======================================================
soundfile==0.12.1
librosa==0.10.0
torchcodec==0.2.1

# ======================================================
# Hugging Face – Models & Datasets
# ======================================================
transformers==4.40.2
datasets==2.14.6
evaluate==0.4.0
sentencepiece==0.1.99

# ======================================================
# ASR Evaluation
# ======================================================
jiwer==3.0.1
pyctcdecode==0.10.0

# ======================================================
# ONNX Ecosystem
# ======================================================
onnx==1.15.0
onnxscript==1.11.0
onnxruntime==1.16.0
onnxruntime-gpu==1.16.0
onnxruntime-tools==1.8.0

# ======================================================
# Model Visualization
# ======================================================
netron==7.7.5

# ======================================================
# Triton Inference Server (Client Side)
# ======================================================
tritonclient==2.20.0

# ======================================================
# Utilities / Developer Experience
# ======================================================
requests==2.32.4
rich==13.4.0
click==8.1.7
python-magic==0.4.27

# ======================================================
# Optional: Notebook Support
# ======================================================
ipython==8.12.3
# jupyterlab==4.0.11

# ======================================================
# IMPORTANT NOTES (DO NOT REMOVE)
# ======================================================
# 1. DO NOT install torch / torchaudio / torchvision from this file.
#    Install them separately using the official PyTorch CUDA command.
#
# 2. TensorRT Python bindings are NOT installed via pip reliably.
#    Use NVIDIA TensorRT system installation for WSL2.
#
# 3. KenLM is optional but recommended for LM decoding:
#    pip install https://github.com/kpu/kenlm/archive/master.zip
